move useLgTemplate to 'hooks' folder
